refactor(analytics-core): move remote config from analytics-browser to analytics-core
