+++ packages/analytics-core/src/index.ts
@@ -180,3 +180,4 @@ export { ExcludeInternalReferrersOptions, EXCLUDE_INTERNAL_REFERRERS_CONDITIONS 
 export { VideoObserver, State as VideoState, type VideoObserverParams } from './observers/video';
 export { EmbeddedVideoPlayer, type Vendor as VideoVendor } from './video-analytics/types';
 export { isChromeExtension, isReactNative } from './utils/environment';
+export { translateRemoteConfigToLocal } from './config/joined-config';
