+++ test-server/autocapture/element-interactions.html
@@ -273,7 +273,7 @@
         import.meta.env.VITE_AMPLITUDE_API_KEY,
         import.meta.env.VITE_AMPLITUDE_USER_ID || 'amplitude-typescript test user',
         {
-          fetchRemoteConfig: false,
+          fetchRemoteConfig: true,
           logLevel: 'debug',
           autocapture: {
             attribution: {
